requirements.txt
anyio==4.13.0
certifi==2026.2.25
dotenv==0.9.9
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
pyaes==1.6.1
pyasn1==0.6.3
python-dotenv==1.2.2
python-telegram-bot==22.7
rsa==4.9.1
Telethon==1.43.1
typing_extensions==4.15.0
